cognara/requirements.txt
gradio==3.47.1
opencv-python-headless==4.8.1.78
Pillow==10.0.0
numpy==1.24.3
scikit-image==0.21.0